Don't include plugin config in module cache
Modules are cached by node and executed only once. This variable was therefore only written to once and accessed from the cached version thereafter.

Any manipulation of the config (eg. by the argument parser) only happened once in the same instance as a side effect and lead to excessive transforms, etc.

Fixes #19